# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for code snippet sharing service",
        "title": "Code Paste Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "codelibrary.skwtr.com",
    "basePath": "/api",
    "paths": {
        "/snippet": {
            "post": {
                "description": "Create a new code snippet with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "snippets"
                ],
                "summary": "Create a new snippet",
                "parameters": [
                    {
                        "description": "Snippet payload",
                        "name": "snippet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateSnippetRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.SnippetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.EmptyResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.EmptyResponse"
                        }
                    }
                }
            }
        },
        "/snippet/{id}": {
            "get": {
                "description": "Retrieve a single public code snippet by its unique ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "snippets"
                ],
                "summary": "Get a snippet by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Snippet UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SnippetResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.EmptyResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.EmptyResponse"
                        }
                    }
                }
            }
        },
        "/snippet/{id}/raw": {
            "get": {
                "description": "Retrieve the raw code content of a public snippet by its unique ID. Returns plain text.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "snippets"
                ],
                "summary": "Get raw snippet code by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Snippet UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Raw code content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.EmptyResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.EmptyResponse"
                        }
                    }
                }
            }
        },
        "/snippets": {
            "get": {
                "description": "Retrieve a paginated list of all public code snippets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "snippets"
                ],
                "summary": "Get all snippets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search term to filter snippets by title, description, or language",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page (default: 20, max: 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SnippetListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.EmptyResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.EmptyResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create multiple code snippets in a single request. Accepts an array of snippet objects.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "snippets"
                ],
                "summary": "Batch create snippets",
                "parameters": [
                    {
                        "description": "Array of snippet payloads",
                        "name": "snippets",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CreateSnippetRequest"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.EmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/snippets/{id}/views": {
            "post": {
                "description": "Increment the view count of a snippet by its ID. Deduplicates by hashing the client IP and snippet ID, so repeated calls from the same IP are ignored.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "snippets"
                ],
                "summary": "Increment snippet view count",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Snippet UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.EmptyResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.EmptyResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.EmptyResponse"
                        }
                    }
                }
            }
        },
        "/stats/languages": {
            "get": {
                "description": "Retrieve a list of unique programming languages used in snippets",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Get unique programming languages",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GlobalLanguagesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.EmptyResponse"
                        }
                    }
                }
            }
        },
        "/stats/snippets": {
            "get": {
                "description": "Retrieve the total number of snippets created",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Get total snippets count",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GlobalSnippetsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.EmptyResponse"
                        }
                    }
                }
            }
        },
        "/stats/views": {
            "get": {
                "description": "Retrieve the sum of views across all snippets",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Get total view count",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GlobalViewsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.EmptyResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CreateSnippetRequest": {
            "type": "object",
            "required": [
                "code",
                "filename",
                "language",
                "title"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "code": {
                    "type": "string",
                    "minLength": 5
                },
                "description": {
                    "type": "string"
                },
                "filename": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "is_public": {
                    "type": "boolean"
                },
                "language": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3
                }
            }
        },
        "models.EmptyResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.GlobalLanguagesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.GlobalSnippetsData": {
            "type": "object",
            "properties": {
                "total_snippets": {
                    "type": "integer"
                }
            }
        },
        "models.GlobalSnippetsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.GlobalSnippetsData"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.GlobalViewsData": {
            "type": "object",
            "properties": {
                "total_views": {
                    "type": "integer"
                }
            }
        },
        "models.GlobalViewsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.GlobalViewsData"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.Meta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.Snippet": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_featured": {
                    "type": "boolean"
                },
                "is_public": {
                    "type": "boolean"
                },
                "language": {
                    "type": "string"
                },
                "snippet_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
        "models.SnippetListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Snippet"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/models.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.SnippetResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Snippet"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}